UPDATE: Now represents 8 ASL Symbols
This is an updated version of the previous file initially representing 5 ASL symbol, now 8 ASL symbols. It has a larger dataset that has been meticulously labelled to ensure that all given input data is correct. Furthermore, it also includes saving the neural network model into JSON format and freezing the graph to store information regarding the architecture of the network and given checkpoints along with the value of weights.